# openapi.json
{"openapi":"3.1.0","info":{"version":"1.0.0","title":"OpenAPI"},"components":{"schemas":{},"parameters":{}},"paths":{"/api/users":{"get":{"operationId":"get_ListUsers","parameters":[{"schema":{"type":"integer","minimum":1,"default":1},"required":false,"name":"page","in":"query"},{"schema":{"type":"integer","minimum":1,"maximum":100,"default":20},"required":false,"name":"per_page","in":"query"}],"responses":{"200":{"description":"List objects","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"result":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"created":{"type":"string","format":"date-time"},"rolename":{"type":"string","minLength":5},"email":{"type":"string","format":"email"},"useruuid":{"type":"string","format":"uuid"}},"required":["id","created","rolename","email","useruuid"]}}},"required":["success","result"]}}}}}},"post":{"operationId":"post_CreateUser","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"}},"required":["email"],"additionalProperties":{}}}}},"responses":{"201":{"description":"Returns the created Object","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"result":{"type":"object","properties":{"email":{"type":"string","format":"email"}},"required":["email"],"additionalProperties":{}}},"required":["success","result"]}}}},"400":{"description":"Input Validation Error","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","enum":[false]},"errors":{"type":"array","items":{"type":"object","properties":{"code":{"type":"number","example":7001},"message":{"type":"string","example":"Input Validation Error"},"path":{"type":"array","items":{"type":"string","example":"body"}}},"required":["code","message","path"]}}},"required":["success","errors"]}}}}}}},"/api/users/{id}":{"get":{"operationId":"get_GetUser","parameters":[{"schema":{"type":"number"},"required":true,"name":"id","in":"path"}],"responses":{"200":{"description":"Returns a single object if found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"result":{"type":"object","properties":{"id":{"type":"number"},"created":{"type":"string","format":"date-time"},"rolename":{"type":"string","minLength":5},"email":{"type":"string","format":"email"},"useruuid":{"type":"string","format":"uuid"}},"required":["id","created","rolename","email","useruuid"]}},"required":["success","result"]}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","enum":[false]},"errors":{"type":"array","items":{"type":"object","properties":{"code":{"type":"number","example":7002},"message":{"type":"string","example":"Not Found"}},"required":["code","message"]}}},"required":["success","errors"]}}}}}}},"/api/lessons":{"get":{"operationId":"get_ListLessons","parameters":[{"schema":{"type":"integer","minimum":1,"default":1},"required":false,"name":"page","in":"query"},{"schema":{"type":"integer","minimum":1,"maximum":100,"default":20},"required":false,"name":"per_page","in":"query"}],"responses":{"200":{"description":"List objects","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"result":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"created":{"type":"string","format":"date-time"},"title":{"type":"string","minLength":3,"maxLength":255}},"required":["id","created","title"]}}},"required":["success","result"]}}}}}},"post":{"operationId":"post_CreateLesson","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","minLength":3,"maxLength":255}},"required":["title"],"additionalProperties":{}}}}},"responses":{"201":{"description":"Returns the created Object","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"result":{"type":"object","properties":{"title":{"type":"string","minLength":3,"maxLength":255}},"required":["title"],"additionalProperties":{}}},"required":["success","result"]}}}},"400":{"description":"Input Validation Error","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","enum":[false]},"errors":{"type":"array","items":{"type":"object","properties":{"code":{"type":"number","example":7001},"message":{"type":"string","example":"Input Validation Error"},"path":{"type":"array","items":{"type":"string","example":"body"}}},"required":["code","message","path"]}}},"required":["success","errors"]}}}}}}},"/api/lessons/{id}":{"get":{"operationId":"get_GetLesson","parameters":[{"schema":{"type":"number"},"required":true,"name":"id","in":"path"}],"responses":{"200":{"description":"Returns a single object if found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"result":{"type":"object","properties":{"id":{"type":"number"},"created":{"type":"string","format":"date-time"},"title":{"type":"string","minLength":3,"maxLength":255}},"required":["id","created","title"]}},"required":["success","result"]}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","enum":[false]},"errors":{"type":"array","items":{"type":"object","properties":{"code":{"type":"number","example":7002},"message":{"type":"string","example":"Not Found"}},"required":["code","message"]}}},"required":["success","errors"]}}}}}}},"/api/courses":{"get":{"operationId":"get_ListCourses","parameters":[{"schema":{"type":"integer","minimum":1,"default":1},"required":false,"name":"page","in":"query"},{"schema":{"type":"integer","minimum":1,"maximum":100,"default":20},"required":false,"name":"per_page","in":"query"}],"responses":{"200":{"description":"List objects","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"result":{"type":"array","items":{"type":"object","properties":{"courseId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"image":{"type":"string","format":"uri"},"subject":{"type":"string"},"instructor":{"type":"string"},"updated":{"type":"string","format":"date-time"},"published":{"type":"string","format":"date-time"}},"required":["courseId","title","description","image","subject","instructor","updated","published"]}}},"required":["success","result"]}}}}}},"post":{"operationId":"post_CreateCourse","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"courseId":{"type":"string","minLength":3}},"required":["courseId"],"additionalProperties":{}}}}},"responses":{"201":{"description":"Returns the created Object","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"result":{"type":"object","properties":{"courseId":{"type":"string","minLength":3}},"required":["courseId"],"additionalProperties":{}}},"required":["success","result"]}}}},"400":{"description":"Input Validation Error","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","enum":[false]},"errors":{"type":"array","items":{"type":"object","properties":{"code":{"type":"number","example":7001},"message":{"type":"string","example":"Input Validation Error"},"path":{"type":"array","items":{"type":"string","example":"body"}}},"required":["code","message","path"]}}},"required":["success","errors"]}}}}}}},"/api/courses/{courseId}":{"get":{"operationId":"get_GetCourse","parameters":[{"schema":{"type":"string"},"required":true,"name":"courseId","in":"path"}],"responses":{"200":{"description":"Returns a single object if found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"result":{"type":"object","properties":{"courseId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"image":{"type":"string","format":"uri"},"subject":{"type":"string"},"instructor":{"type":"string"},"updated":{"type":"string","format":"date-time"},"published":{"type":"string","format":"date-time"}},"required":["courseId","title","description","image","subject","instructor","updated","published"]}},"required":["success","result"]}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","enum":[false]},"errors":{"type":"array","items":{"type":"object","properties":{"code":{"type":"number","example":7002},"message":{"type":"string","example":"Not Found"}},"required":["code","message"]}}},"required":["success","errors"]}}}}}}}},"webhooks":{}}